requirements.txt
json
urllib
pyyaml